Made some changes to ensure 64-bit builds work nicely.

+++ Digipost/NSString+SHA1String.m
@@ -20,7 +20,7 @@
 
     uint8_t digest[CC_SHA1_DIGEST_LENGTH];
 
-    CC_SHA1(data.bytes, data.length, digest);
+    CC_SHA1(data.bytes, (unsigned int)data.length, digest);
 
     NSMutableString *SHA1StringMutable = [NSMutableString stringWithCapacity:CC_SHA1_DIGEST_LENGTH * 2];
